Adding AtomicCounter, ProgressBar, Logger
AtomicCounter - Using AtomicCounter to keep track of the number of objects processed so far. This is then used for displaying ProgressBar
ProgressBar - Using for showing visual progress to the user. It is updated everytime when an object is processed.
Logger - Contains functions for logging info/debug/error messages to a log file